Ikke bruk getInformasjonOmAGP2HvisAGP2IkkeNås lenger
