+++ src/komponenter/kalkulator/Tidslinje/tidslinjefunksjoner.tsx
@@ -9,8 +9,9 @@ import {
     formaterDato,
 } from '../utils/dato-utils';
 import {
+    finn18mndsperiodeForMaksimeringAvPermitteringsdager,
     finnDatoAGP2EtterInnføringsdato,
-    finnPermitteringssituasjon, getInformasjonOmAGP2HvisAGP2IkkeNås,
+    finnPermitteringssituasjon,
     Permitteringssituasjon,
 } from '../utils/beregningerForAGP2';
 
@@ -206,13 +207,13 @@ export const finnSisteDatoI18mndsintervalletSomMarkeresITidslinjen = (
     tidslinje: DatoMedKategori[],
     innføringsdatoAGP2: Dayjs,
     antallDagerFørAGP2Inntreffer: number,
-    dagensDato: Dayjs,
+    dagensDato: Dayjs
 ): Dayjs => {
     const situasjon = finnPermitteringssituasjon(
         tidslinje,
         innføringsdatoAGP2,
         antallDagerFørAGP2Inntreffer
-    )
+    );
 
     let sluttDatoIllustrasjonPåTidslinje: Dayjs | undefined;
     switch (situasjon) {
@@ -221,15 +222,15 @@ export const finnSisteDatoI18mndsintervalletSomMarkeresITidslinjen = (
                 tidslinje,
                 innføringsdatoAGP2,
                 antallDagerFørAGP2Inntreffer
-            )
+            );
             break;
         case Permitteringssituasjon.AGP2_IKKE_NÅDD_PGA_FOR_LITE_PERMITTERT:
-            sluttDatoIllustrasjonPåTidslinje = getInformasjonOmAGP2HvisAGP2IkkeNås(
+            sluttDatoIllustrasjonPåTidslinje = finn18mndsperiodeForMaksimeringAvPermitteringsdager(
                 tidslinje,
                 innføringsdatoAGP2,
-                antallDagerFørAGP2Inntreffer,
-                dagensDato
-            ).sluttDato;
+                dagensDato,
+                antallDagerFørAGP2Inntreffer
+            )?.datoTil;
             break;
         case Permitteringssituasjon.AGP2_IKKE_NÅDD_PGA_IKKE_PERMITTERT_VED_INNFØRINGSDATO:
         case Permitteringssituasjon.AGP2_NÅDD_VED_INNFØRINGSDATO:
@@ -238,4 +239,4 @@ export const finnSisteDatoI18mndsintervalletSomMarkeresITidslinjen = (
             break;
     }
     return sluttDatoIllustrasjonPåTidslinje || innføringsdatoAGP2;
-}
+};
